Added truncate method to projection definition to allow in-place rebuild
