+++ src/packages/emmett-postgresql/src/eventStore/projections/pongo/index.ts
@@ -1,2 +1,2 @@
+export * from './pongoProjections';
 export * from './pongoProjectionSpec';
-export * from './projections';
